task finished

+++ Public/JS/deleteProducts.js
@@ -1,4 +1,4 @@
-$('#delete-product-ptn').on('click',function() {
+$('#delete-product-btn').on('click',function() {
 
     var skus = {skus:[]};
     var counter = 0;
@@ -7,7 +7,7 @@ $('#delete-product-ptn').on('click',function() {
     checkboxes = Array.from($checkboxes);
 
     checkboxes.forEach(function(checkbox){
-
+        
         if(checkbox.checked == true) {
             skus.skus[counter] = $(checkbox).attr('sku');
             counter++;
@@ -15,25 +15,31 @@ $('#delete-product-ptn').on('click',function() {
 
     })
 
-    $.ajax({
+    if((skus.skus).length > 0) {
+        $.ajax({
 
-        method : 'get',
-        url : 'deleteAll',
-        data : skus,
+            method : 'get',
+            url : 'deleteAll',
+            data : skus,
 
-        success:function(data) {
-            console.log(data);
-            if(data.status == true) {
-                checkboxes.forEach(function(checkbox) {
-                    if(checkbox.checked)
-                        checkbox.parentNode.remove();
-                });
+            success:function(data) {
+            
+                if(data.status == true) {
+                    checkboxes.forEach(function(checkbox) {
+                        if(checkbox.checked)
+                            checkbox.parentNode.remove();
+                    });
+                }
+
+                if(checkboxes.length == 0) {
+                    $('#delete-product-btn').hide(300);
+                }
+            },
+            error : function(xhr,status,error) {
+                alert('error');
             }
-        },
-        error : function(xhr,status,error) {
-            alert('error');
-        }
 
-    })
+        });
+    }
 
 })
